Writer class implementation

+++ src/main/java/ru/hfm/convertor/Writer.java
@@ -7,10 +7,7 @@ package ru.hfm.convertor;
  * @since 06.04.2017
  */
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
 import java.util.Iterator;
 import java.util.List;
 
@@ -20,6 +17,7 @@ class Writer {
     private File file;
     private List<DataRecord> dataArray;
     private StringBuilder resultString;
+    private final String DEFAULT_DEPARATOR = ";";
 
     Writer(Parameters parameters, List<DataRecord> dataArray) {
         this.parameters = parameters;
@@ -31,59 +29,77 @@ class Writer {
         this.file = new File(parameters.getOutputFile());
     }
 
-    void writeFile() {
+    private void createNewFile() {
 
         if (!this.file.exists()) {
             try {
                 this.file.createNewFile();
-            } catch (Exception exception) {
-                exception.printStackTrace();
+            } catch (IOException ioException) {
+                ioException.printStackTrace();
             }
         } else if (this.file.exists()) {
-            this.file.delete();
+            if (this.file.canWrite()) {
+                this.file.delete();
+            }
         }
+
+    }
+
+    void writeFile() {
+
+        createNewFile();
 
         try (FileWriter fileWriter = new FileWriter(this.file.getAbsoluteFile())) {
 
             try (BufferedWriter bufferedWriter = new BufferedWriter(fileWriter, 1024 * 1024)) {
 
-                Iterator<DataRecord> dataArrayIterator  = dataArray.iterator();
+                Iterator<DataRecord> dataArrayIterator = dataArray.iterator();
 
                 while (dataArrayIterator.hasNext()) {
 
                     DataRecord dataRecord = dataArrayIterator.next();
-                    String resultStringLine = constructLine(dataRecord);
+                    StringBuilder resultStringLine = constructLine(dataRecord);
+                    bufferedWriter.write(resultStringLine.toString());
+                    this.resultString.delete(0, resultString.length());
                     bufferedWriter.newLine();
-
 
                 }
 
-            } catch (IOException exception) {
-                exception.printStackTrace();
+                bufferedWriter.flush();
+
+                bufferedWriter.close();
+
+            } catch (IOException ioException) {
+                ioException.printStackTrace();
             }
 
-        } catch (IOException exception) {
-            exception.printStackTrace();
+        } catch (IOException ioException) {
+            ioException.printStackTrace();
         }
 
 
     }
 
-    private String constructLine(DataRecord dataRecord) {
+    private StringBuilder constructLine(DataRecord dataRecord) {
 
         this.resultString.append(dataRecord.getSourceFMEntity());
+        this.resultString.append(DEFAULT_DEPARATOR);
         this.resultString.append(dataRecord.getSourceFMAccount());
+        this.resultString.append(DEFAULT_DEPARATOR);
         this.resultString.append(dataRecord.getSourceICP());
+        this.resultString.append(DEFAULT_DEPARATOR);
         this.resultString.append(dataRecord.getSourceCustom1());
+        this.resultString.append(DEFAULT_DEPARATOR);
         this.resultString.append(dataRecord.getSourceCustom2());
+        this.resultString.append(DEFAULT_DEPARATOR);
         this.resultString.append(dataRecord.getSourceCustom3());
+        this.resultString.append(DEFAULT_DEPARATOR);
         this.resultString.append(dataRecord.getSourceCustom4());
+        this.resultString.append(DEFAULT_DEPARATOR);
         this.resultString.append(dataRecord.getAmount());
 
-        String resultStringLine = new String(resultString);
-        resultString.delete(0, resultString.length() - 1);
+        return resultString;
 
-        return resultStringLine;
     }
 
 }
